SVN r23874 |2013-01-07 15:12:28 +0000

+++ ehr/resources/queries/ehr/protocolTotalAnimalsBySpecies.sql
@@ -11,7 +11,7 @@ p.species,
 pc.allowed,
 p.TotalAnimals,
 CONVERT(pc.allowed - p.TotalAnimals, INTEGER) as TotalRemaining,
-round(cast(p.TotalAnimals as numeric) / cast(pc.allowed as numeric) * 100, 1) as PercentUsed,
+round(cast(p.TotalAnimals as float) / cast(pc.allowed as float) * 100, 1) as PercentUsed,
 p.Animals,
 
 FROM
@@ -45,7 +45,7 @@ SELECT
   p.maxAnimals as allowed,
   CONVERT(Count(pa.id), INTEGER) AS TotalAnimals,
   p.maxAnimals - Count(pa.id) as TotalRemaining,
-  round(cast(Count(pa.id) as numeric) / cast(p.maxAnimals as numeric) * 100, 1) as PercentUsed,
+  round(cast(Count(pa.id) as float) / cast(p.maxAnimals as float) * 100, 1) as PercentUsed,
   group_concat(distinct pa.Id) as animals,
 
 FROM ehr.protocol p
