+++ ehr/resources/queries/study/majorEvents.sql
@@ -11,7 +11,7 @@ b.date,
 case
   when b.id.dataset.birth.weight is null
     then ('Offspring: ' || b.id)
-  else ('Offspring: ' || b.id || chr(10) || 'Offspring Weight: ' || cast(cast(b.id.dataset.birth.weight as numeric) as varchar)|| ' kg')
+  else ('Offspring: ' || b.id || chr(10) || 'Offspring Weight: ' || cast(cast(b.id.dataset.birth.weight as float) as varchar)|| ' kg')
 END as remark
 from study.birth b
 WHERE b.qcstate.publicdata = true AND b.dam is not null and b.dam != ''
